requirement.txt
python-dotenv
langchain
langchain_community
langchain_openai
openai
faiss-cpu
PyPDF2
streamlit
-e .